Move *Metrics classes into their own files
